+++ src/support/utils.js
@@ -63,6 +63,19 @@ export const sendExperimentationCandidatesMessage = async (
 });
 /* c8 ignore end */
 
+export const sentRunScraperMessage = async (
+  sqs,
+  queueUrl,
+  jobId,
+  urls,
+  slackContext,
+) => sqs.sendMessage(queueUrl, {
+  processingType: 'default',
+  jobId,
+  urls: [...urls],
+  slackContext,
+});
+
 // todo: prototype - untested
 /* c8 ignore start */
 export const sendRunImportMessage = async (
@@ -149,6 +162,21 @@ export const triggerExperimentationCandidates = async (
 );
 /* c8 ignore end */
 
+export const triggerScraperRun = async (
+  jobId,
+  urls,
+  slackContext,
+  lambdaContext,
+) => sentRunScraperMessage(
+  lambdaContext.sqs,
+  lambdaContext.env.SCRAPING_JOBS_QUEUE_URL,
+  jobId,
+  urls,
+  {
+    channelId: slackContext.channelId,
+    threadTs: slackContext.threadTs,
+  },
+);
 // todo: prototype - untested
 /* c8 ignore start */
 export const triggerImportRun = async (